typescript + setup radix
